refactor(storage): move the multipart machinery into s3_multipart.rs
Pure code motion, requested in review: the part writer, uploader thread,
dedicated-client builder and the s3_put_versioned/s3_finish_put pair move out
of s3.rs (which was past 2300 lines) into their own module, together with the
multipart constants and the writer's unit tests. s3.rs keeps the S3Storage
trait impls, the single-PUT helpers, the mock and the storage-level tests.
No behavior change.
